Revert "Move lsp module to command runners module."
This reverts commit 89eecc3b1bf6617652d77799c3673eaec3be33e7.
