+++ src/pre_commit_sbt/lsp/conn.py
