+++ src/pre_commit_sbt/lsp/rpc.py
