+++ src/pre_commit_sbt/lsp/server.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from pre_commit_sbt.command_runners.lsp.port_file import port_path
+from pre_commit_sbt.lsp.port_file import port_path
 
 
 def is_server_running(root_dir: Path) -> bool:
